- made some changes to memory stuff

+++ Minecraft Clone/src/Core/Options/Option.h
@@ -6,14 +6,14 @@
 
 class Options {
 public:
-    int vertical_render_distance_ = 8;
-    int horizontal_render_distance_ = 16;
-    int world_gen_threads_ = 6;
-    int mesh_threads_ = 6;
-    int graphics_scale_ = 2;
-    int transparent_buffer_size_ = 1000000000;
-    int solid_buffer_size_ = 2000000000;
-    int light_engine_threads_ = 2;
+    size_t vertical_render_distance_ = 8;
+    size_t horizontal_render_distance_ = 16;
+    size_t world_gen_threads_ = 6;
+    size_t mesh_threads_ = 6;
+    size_t graphics_scale_ = 2;
+    size_t transparent_buffer_size_ = 1000000000;
+    size_t solid_buffer_size_ = 2000000000;
+    size_t light_engine_threads_ = 2;
 
     Options();
 private:
@@ -26,6 +26,6 @@ private:
 
     void GenerateOptionFile(); //Generate file if deleted
 
-    FastHashMap<std::string, int*> option_name_; //name -> ptr
+    FastHashMap<std::string, size_t*> option_name_; //name -> ptr
 
 } extern g_app_options;
